renaming av filer og mapper, slett ubrukt css
