+++ src/components/_common/chatbot/ChatbotLinkPanel.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { openChatbot } from '@navikt/nav-dekoratoren-moduler';
-import { LinkPanelNavno } from 'components/_common/linkpanel/LinkPanelNavno';
+import { LinkPanelNavno } from 'components/_common/linkpanel/LinkPanelNavno/LinkPanelNavno';
 import { FrontpageContactAlert } from 'components/parts/frontpage-contact/FrontpageContactAlert';
 import { ParsedHtml } from 'components/_common/parsed-html/ParsedHtml';
