+++ src/components/_common/linkpanel/LinkPanelNavnoSimple/LinkPanelNavnoSimple.module.scss
