+++ src/components/_common/linkpanel/LinkPanelNavnoSimple/LinkPanelNavnoSimple.stories.tsx
