+++ src/components/_common/linkpanel/LinkPanelNavnoSimple/LinkPanelNavnoSimple.tsx
