+++ src/components/parts/frontpage-contact/FrontpageContactPart.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Heading } from '@navikt/ds-react';
 import { EditorHelp } from 'components/_editor-only/editor-help/EditorHelp';
-import { LinkPanelNavno } from 'components/_common/linkpanel/LinkPanelNavno';
+import { LinkPanelNavno } from 'components/_common/linkpanel/LinkPanelNavno/LinkPanelNavno';
 import { ContentType } from 'types/content-props/_content-common';
 import { ChatbotLinkPanel } from 'components/_common/chatbot/ChatbotLinkPanel';
 import { PartComponentProps, PartType } from 'types/component-props/parts';
