+++ src/components/parts/linkPanelPart/LinkPanelPart.module.scss
